+++ src/Domain/Services/IScheduleService.cs
@@ -5,7 +5,7 @@ namespace GridMonitor.Domain.Services;
 
 public interface IScheduleService
 {
-	ValueTask<Result<int>> GetCurrentStageAsync(CancellationToken ct = default);
+	ValueTask<Result<short>> GetCurrentStageAsync(CancellationToken ct = default);
 	ValueTask<Result<SuburbSchedule>> GetScheduleAsync(int suburbId, CancellationToken ct = default);
 	ValueTask<Result<List<ScheduleSlot>>> GetUpcomingAsync(int suburbId, int currentStage, CancellationToken ct = default);
 	ValueTask<Result<List<Suburb>>> SearchSuburbsAsync(string query, CancellationToken ct = default);
